% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_funtions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots the acidents on a map for a user specified year and state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{interger specifying the state}

\item{year}{interger specifying the year of data to be ploted}
}
\value{
The function plots the acidents for a given year
on a map of the state
}
\description{
User specifies year and state (as function parameters) and the function
will plot the acidents for that given year, on a map of the state
Acidents will be ploted has dots, based on their latitude and longitude
}
\examples{
\dontrun{
fars_map_state (45,2014)
}

}
