% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_funtions.R
\name{fars_read}
\alias{fars_read}
\title{Reading data from file
Funtions reads data from Fatality Analysis Reporting System
The function will read the defined data file into data table.
If the data file is not available it will give an error message}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{string indicating the data name}
}
\value{
The function will read the file extablishe in the function parameter, and loads it in to a data table
If file does not exist, the function will return an error message.
}
\description{
Reading data from file
Funtions reads data from Fatality Analysis Reporting System
The function will read the defined data file into data table.
If the data file is not available it will give an error message
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz")
}

}
