% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_funtions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads several years of data from the data set}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a list of integers representing the years}
}
\value{
The function a data table for each year
for year not present in the data, the function gives a warning message
}
\description{
The function takes the years specified by the user has a parameter and
creates a table for each year. Each table is a list with two collumns (year, month)
}
\examples{
\dontrun{
fars_read_years(2013:2015)
}

}
