% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_funtions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Outputs a summary of the years and months in the data set}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a list of integers representing the years}
}
\value{
The function a prints out a summary table of the observations
grouped by year/month
}
\description{
Based on the years provided by the user has a parameter,
the funcion prints out the summary of observations grouped by years and months
}
\examples{
\dontrun{
fars_summarize_years (2013:2015)
}

}
