% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_funtions.R
\name{make_filename}
\alias{make_filename}
\title{Generate filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{an integer}
}
\value{
The function generates a filename with the year number given by the user.
}
\description{
The function generates a "accident_year.csv.bz2" filename,
where "year" is the parameter given by the user.
}
\examples{
\dontrun{
make_filename(2013)
}

}
