% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot all accident locations for a given year on a US federal state map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Number representing a US federal state.}

\item{year}{Input vector specifying a single year for which the accidents should be plotted.}
}
\description{
\code{fars_map_state()} consumes the informaton stored in a single FARS report
and creates a plot showing the location of all accidents for a specified US
federal state and year.
}
\details{
This function has no return value. It is only used for its plotting side effects.
A message will be printed to the console if no accidents were reported for a given
and valid state-year combination. \code{fars_map_state()} expects to find
the report file in the current working directory.
}
\examples{
\dontrun{
fars_map_state(13, 2014)
}
}
