% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot FARS fatalities for a state and year.}
\usage{
fars_map_state(state.num, year, path = "inst/extdata")
}
\arguments{
\item{state.num}{The integer number of the state to be plotted, from 1-56.}

\item{year}{The year to be plotted}

\item{path}{The path to the FARS data file to read. The default is the path
to the three supplied raw data files.}
}
\value{
A maps object.
}
\description{
Makes a plot of Fatality Analysis Reporting System (FARS)
data for a given state number and year.
}
\examples{
\dontrun{fars_map_state(state.num = 10, year = 2014)}     # Returns a map for state 10.
\dontrun{fars_map_state(state.num = 100, year = 2014)}     # Returns an error.

}
