% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Visualize Accidents within a state of the US on the map.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The state number for which we want to plot accidents on the map.}

\item{year}{The year for which we want to look at accidents on the map.}
}
\value{
A map with accidents plotted on the map for the selected year and selected state number.
}
\description{
This is a simple function that allows us to show accidents by exact location in a state within a year on a map.
If there are no accidents for a state in a year,
then a message "no accidents to plot" would vbe shown.
The "state.num" should be valid otherwise the function would stop
and show the message "invalid STATE number: 'state.num'".
}
