% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{the number or string representing the state ID in dataset STATE}

\item{year}{a number or string of the year}
}
\value{
If there are no accidents to plot, then a message is returned. Otherwise, a map
of the state with accidents is generated
}
\description{
The purpose of this function is to provide a state-wide map of accidents for a \code{year}
provided by the userm, for the state with code \code{state.num}. It plots the position of each accident using long and lat
data provided by the data set "STATE". The function handles invalid state numbers
and cases where there are no accidents to report.
}
\examples{
\dontrun{fars_map_state("1","2013")}

}
