% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accidents on a map.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer identifying the state.}

\item{year}{The year of interest.}
}
\value{
A plot with the state outline and individual points representing accidents.
}
\description{
This function will plot accidents on a map, for a given state and year. Each dot represents a single accident.
It will fail if the user provides an invalid \code{state.num} or \code{year}.
}
