% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draw a state map with dots showing location of each accident}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Numeric state number. If \code{state.num} cannot be converted to an integer or there is no
state corresponding to this number an error will be generated}

\item{year}{Numeric year for which to draw a map}
}
\description{
This function generates a picture of the state map with dots showing location of each accident
for the given year and state
}
\note{
This function has no return value. It is used only for its side effect of drawing the picture.
      If there are no recorded accidents for a given state-year combination a message will be reported indicating this.
}
\examples{
\dontrun{
fars_map_state(6,2015)
}

}
