% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Visualising Incidents of selected state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{an integer representing the State Number}

\item{year}{Indicating the year of interest. It should be entered as 4 digit
integer.}
}
\value{
None.
}
\description{
The function plots incidents of selected state and year from the FARS
dataset. It has exception handling to ensure\itemize{\item the file name is
valid (refer \code{\link{fars_read}}), \item the state number does exist in
the data set, and \item there is indeed data for the selected state number.}
Incidents that has longitude over 900 and Latitude over 90 are discarded.
}
\examples{
\dontrun{fars_map_state(1, 2015)}

}
\seealso{
\code{\link{fars_read}} for reading file
}
