% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Name of the file to read}
}
\value{
This function returns a data frame containing the data read from the file
if the requested file does not exist, an error will be reported and the function
will stop
}
\description{
A FARS file contains data from the US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System, which is a nationwide
census providing the American public yearly data regarding fatal injuries
suffered in motor vehicle traffic crashes.
}
\details{
Requires dpyr, tidyr, readr, maps, graphics
}
\examples{
\dontrun{
fars_read("data/accident_2015.csv.bz2")
}

}

