% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fars_read}
\alias{fars_read}
\title{Read a compressed csv report file downloaded from the Fatality
Analysis Reporting System (FARS) into a tibble}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path to a file containing FARS data.}
}
\value{
A tibble. If the file cannot be found at the specified path,
an error is thrown.
}
\description{
\code{fars_read()} reads a locally stored and compressed csv file downloaded from the
\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{Fatality Analysis Reporting System}
into a tibble.
}
\examples{
\dontrun{
fars_complete_2013 <- fars_read("accident_2013-csv.bz2")
}
}
