% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read Fatality Analysis Reporting System data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the FARS data file to read.}

\item{path}{The path to the FARS data file to read. The default is the path
to the three supplied raw data files.}
}
\value{
A table of FARS data.
}
\description{
\code{fars_read} reads in Fatality Analysis Reporting System (FARS) data
for a given \code{filename}, if the file exists.
}
\examples{
\dontrun{fars_read(filename = "accident_2013.csv.bz2")}
\dontrun{fars_read(filename = "accident_2014.csv.bz2"))}
\dontrun{fars_read(filename = "accident_2015.csv.bz2")}
\dontrun{fars_read(filename = "file_does_not_exist.csv", path = "bad/path")} # Results in an error

}
