% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read Fatality Analysis Reporting System data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name (and path to) the FARS data file to read.}
}
\value{
A table of FARS data.
}
\description{
\code{fars_read} reads in Fatality Analysis Reporting System (FARS) data
for a given \code{filename}, if the file exists.
}
\examples{
fars_read(filename = "data/accident_2013.csv.bz2")
fars_read(filename = "data/accident_2014.csv.bz2")
fars_read(filename = "data/accident_2015.csv.bz2")
\dontrun{fars_read(filename = "file_does_not_exist.csv")} # Results in an error

}
