% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far_function.R
\name{fars_read}
\alias{fars_read}
\title{Read data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string of the filename of the data that need to be read}
}
\value{
a dataframe containing a representation of the data with class of tibble
}
\description{
This is a simple function that reads a file by taking filename as input.
If a filename doesn't exist in the working directory then the function
stops with the message "file 'filename' does not exist"
If a filename exist, the file will be read in this function as an imported data
}
\examples{
fars_read("accident_2013.csv.bz2")
data <- fars_read("accident_2014.csv.bz2")
}
