% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Read csv file.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string containing the filename.}
}
\value{
A tibble object containing data from the corresponding file.
}
\description{
This function will read a csv file into a tibble. It will fail to run
if the user provides an invalid filename.
}
