% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS Data File}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string or object representing path to files}
}
\value{
A tibble contains the contents of input file, the first part of which will
  be print out to console as a side effect. Invalid path will result in a "file ...
  does not exist" error.
}
\description{
`fars_read()` is a helper function used by `fars_read_years()` and `fars_map_state()`,
which reads in an annual FARS data file. The file is supposedly the query output from
the US National Highway Traffic Safety Administration's Fatality Analysis Reporting
System, exported into a .csv file and compressed in the bz2 format.
}
\examples{
\dontrun{
fars_read("accident_2013.csv")
}
}
