% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read \url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{Fatality Analysis Reporting System (FARS)} report file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Name of the report file to read. An error will be thrown if the \code{filename} does not exist}
}
\value{
This function returns a data frame tibble containing the information from the report file
}
\description{
This function reads a FARS report file in CSV format and returns a data frame tibble
}
\examples{
\dontrun{
df <- fars_read("../data/accident_2013.csv.bz2")
}

}
