% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS data from a file into R}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string specifying the file name of the FARS data}
}
\value{
a tibble representation of the data
}
\description{
The function acessess and reads in data from the given filename. It
throws an error if the file does not exist. The data is then represented as a
tibble. Note that any warning and progress messages are suppressed during the
read process.
}
\examples{
\dontrun{fars_read('accident_2015.csv.bz2')}

}
