% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads months and years from multiple FARS files (for multiple years)}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector or list of years for which to read FARS files}
}
\value{
This function returns a list of months and years of accidents from the specified datafile(s)
If any error occurs, a warning is generated: "invalid year: XXXX"
}
\description{
Notes:
The FARS file(s) must be located in the current working directory.
}
\examples{
\dontrun{
fars_read_years(2013:2015)
fars_read_years(list(2013, 2015))
}

}

