% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read specific information of multiple FARS report files into a list of tibbles}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Input vector or list specifying one or several years in character or
integer format.}
}
\value{
A list of tibbles. Each tibble contains just two columns of the original
FARS data set (\code{MONTH} and \code{year}). Each row corresponds to a single accident.
}
\description{
\code{fars_read_years()} reads several locally stored and compressed FARS report
 files for the specified years into a list of tibbles.
}
\details{
Only the monthly and yearly information of each accident included in an
individual FARS report is stored in the respective tibble reducing each original
data set to two columns. \code{fars_read_years()} expects to find the
report files in the current working directory.
}
\examples{
\dontrun{
fars2013 <- fars_read_years(2013)
fars2013 <- fars_read_years("2013")
fars1314 <- fars_read_years(c(2013, 2014))
}
}
