% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read the data from multiple files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
A list of data frames.
}
\description{
This is a simple function that allows us to read multiple files
by inputting the "years" for which we want to read files.
It generates a list of data frames,
where each file is a data frame that's part of the list.
}
