% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector or list of years. These could be either strings or numbers}
}
\value{
Dataframes with columns "Month" and "Year" for each year in the list "years". The
function will retrn a warning if a year in the list is invalid.
}
\description{
Purpose of this function is to create a dataframe that shows the accidents occuring
in each month for each \code{years} provided by the user. This function takes creates a dataframe for each year in a list of years provided
by the user. It takes a list of \code{years}, then uses functions make_filename, and fars_read
to first generate a filename for a specific year, then create a dataframe for the file
with that filename. Lastly, the dataframe has a "year" column added to it, and the
columns "Month" and "Year" are selected for return.
}
\examples{
\dontrun{fars_read_years(c("2013","2014","2015"))}


}
