% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Number of accidents from multiple years.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector containing the years of interest.}
}
\value{
A list of tibbles.
}
\description{
This function will read in data from multiple years and save it in a list,
in which each element is a tibble for each year and month, where each row is
an accident. It will print an error message, if the user provides an invalid year.
}
