% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in Annual FARS data for Multiple Years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a numeric vector of four digit numbers represents years
in the format 'YYYY'}
}
\value{
A list of tibbles containing annual accident entries that corresponds
  to years specified in the argument. An invalid input value (year) will cause
  a NULL value to be returned to the list and an "invalid year..." warning.
}
\description{
`fars_read_years()` is a helper function used by `fars_summarize_years()`,
which reads in fetal traffic accident entries for specified years from
accessible annual FARS data files within the current working directory
}
\examples{
fars_read_years(c(2013,2014,2015))

yrs <- 2013:2015
dat_list <- fars_read_years(years=yrs)
}
\seealso{
\code{\link{fars_read}} and \code{\link{make_filename}},
  which this function wraps
}
