% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple report files for the list of years provided}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Vector or list containing years for which to read the data from the FARS reports}
}
\value{
List of tibbles. Each tibble contains information about \code{MONTH} and \code{year} for each processed measurement
}
\description{
This function reads the data for multiple years provided as a list and returns a list of data frame tibbles
Each tibble will have two columns (variables)
\itemize{
   item MONTH
   item year
}
These columns conorrespond to month and year of each measurement
}
\note{
\code{fars_read_years} uses \code{\link{make_filename}} function to generate the list of files to process.
      \code{fars_read_years} expexts to find the report files in the current working directory
      If one of the report files does not exist \code{fars_read_years} will throw a warning.
}
\examples{
\dontrun{
mon_year <- fars_read_years(c(2013,2014,2015))
}

}
