% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read data from multiple files into R}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list or vector specifying the years of interest}
}
\value{
a list containing data read from the files. Each element corresponds
  to one particular year specified by the file name.
}
\description{
The function reads in data from filenames created from the input
argument. \code{\link{make_filename}} is called to create the filenames. For
each of the filenames, an attempt is made to read the file, and only column
/code{MONTH} and /code{year} are selected and stored as element in a list; if
the operations fail, the list element corresponding to the file will have no
records. Once all files are read, the list is returned.
}
\examples{
\dontrun{fars_read_years(c(2013:2015))}

}
\seealso{
\code{\link{fars_read}} for reading files, and
  \code{\link{make_filename}} for creating the filename
}
