% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Reads months and years from FARS files and summarizes how many accidents occurred per month/year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector or list of years for which to read FARS files}
}
\value{
This function returns a tibble containing the number of accidents per month. Each column represents a year and each row a month.
}
\description{
Notes:
The FARS file(s) must be located in the current working directory.
}
\examples{
\dontrun{
fars_summarize_years(2013:2015)
}

}

