% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS report accidents by month and year in a tibble}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Input vector or list specifying one or several years in character or
integer format.}
}
\value{
A tibble conataining the summarized FARS accident information.
}
\description{
\code{fars_read_years()} creates a summary of several FARS reports showing
monthly accidents by year and stores the result in a single tibble.
}
\details{
\code{fars_summarize_years()} expects to find the report files in the
current working directory.
}
\examples{
\dontrun{
fars_summarize_years(c(2013, 2014, 2015))
}
}
