% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{FARS Monthly Fatality Summaries}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The year or years to get monthly summaries of fatalities for.}
}
\value{
A summary table of monthly fatalities for each year of FARS data.
}
\description{
Creates summaries of monthly fatalities using Fatality Analysis
Reporting System data for a specified year or years.
}
\examples{
\dontrun{fars_summarize_years(years = 2014)}   # Monthly fatality summaries for 2014.
years <- c(2013, 2014, 2015)
\dontrun{fars_summarize_years(years = years)}     # Summary table for 2013-2015.
\dontrun{fars_summarize_years(years = 14)}        # Will return an error.

}
