% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Create Summary by month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
A data frame summarising number of accidents grouped by month and year.
}
\description{
This is a simple function for calculating
a summary of total accidents grouped by month and year.
}
