% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector or list of years. These could be either strings or numbers}
}
\value{
a dataframe that summarizes the number of accidents in a range of years.
Dataframe has two columns: "year" and "n" where "n" represents the number
of accidents in that year.
}
\description{
This function will summarize the accidents in year for a set of \code{years} provided
by the user. It takes a list of \code{years}, makes dataframes for each year using the
"fars_read_years" function. Then using functions from dplyr, the dataframes are joined
grouped by year, and Month. The number of each month / year is counted. Finally, the
dataframe is spread by year and the counts for that year.
}
\examples{
\dontrun{fars_summarize_years(c("2013","2014","2015"))}


}
