% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Number of accidents per year and month.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector with the years of interest.}
}
\value{
A tibble with the number of accidents for the whole country, for each month per year in \code{years}.
}
\description{
This function will summarize the number of accidents for each month in each year of \code{years}.
If the user provides an invalid year, an error message will be printed.
}
