% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Count FARS Traffic Accidents by Month and Year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a numeric vector of four digit numbers represents years
in the format 'YYYY'}
}
\value{
a summary or pivot table of monthly total values, generated by pivoting a
  tibble wider, with rows by months and columns by years; An invalid value (year)
  specified in the input will cause a NULL value generated while reading data from
  corresponding file, missing value (column) for the invalid year in the output,
  and a warning message "invalid year: ...". Completely invalid input will cause a
  "Must group by variables found in `.data`" error in addition to warning messages
  of "invalid year: ..."
}
\description{
`fars_summarize_years()` produces a summary statistic of the fatal
traffic crash data compiled by the US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System (FARS). The
function lets you count monthly totals of fatal traffic accidents
on an annual basis.
}
\details{
\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{See here}
for more information about FARS|NHTSA.
}
\examples{
\dontrun{
fars_summarize_years(c(2013,2014,2015))
}
\dontrun{
yrs <- 2013:2015
monthly_counts <- fars_summarize_years(years=yrs)
}

}
\seealso{
\code{\link{fars_read_years}} which this function wraps
}
