% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summary of the FARS data by Year and Month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list or vector specifying the years of interest}
}
\value{
a data frame showing number of data points per MONTH and year. Rows
  correspond to MONTH while columns correspond to year
}
\description{
The function shows number of data points per month and year in a tabular
format. The summarising operation is done in the following steps: \enumerate{
\item data is specified and read from file, refer
\code{\link{fars_read_years}}, \item data of each year is combined, \item
data is grouped by month and year, \item data is then counted and tabulated }
}
\examples{
\dontrun{fars_summarise_years(c(2013:2015))}

}
\seealso{
\code{\link{fars_read_year}} for reading files, and
  \code{\link{make_filename}} for creating the filename
}
