% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a FARS filename for a particular year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year for which to generate the filename}
}
\value{
This function returns the formatted file name.
If year is not numeric, a warning is generated: "NAs introduced by coercion"
}
\description{
Create a FARS filename for a particular year
}
\examples{
\dontrun{
make_filename("2015")
}

}

