% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_filename}
\alias{make_filename}
\title{Create valid file name for a compressed FARS report for a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Input vector specifying a single year.}
}
\value{
A character vector containing the complete file name of a compressed
FARS report.
}
\description{
Create valid file name for a compressed FARS report for a given year
}
\examples{
make_filename(2013)
}
