% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a Fatality Analysis Reporting System file name.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year you want to make a FARS filename for.}
}
\value{
A filename for a file containing FARS data.
}
\description{
\code{make_filename} creates a Fatality Analysis Reporting System
(FARS) filename for a given four digit \code{year}.
}
\examples{
\dontrun{make_filename(year = 2014)}   # Makes FARS filename for the year 2014.
\dontrun{make_filename(year = 14)}     # Will not make a working FARS filename.

}
