% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A string or number}
}
\value{
The function returns a string with format: "accident_\%d.csv.bz2"
}
\description{
The purpose of this function is to create a filename for the accidents occuring in
a \code{year} given by the user. This function will create a string with
format "accident_\%d.csv.bz2" , where \%d is a year number provided by the user.
There is no error handling.
}
\examples{
\dontrun{make_filename(2013)}

}
