% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make filename for specific \code{year}.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year from which you want data.}
}
\value{
A string with the filename.
}
\description{
This function will create a string with the appropriate filename,
in order to import from a given year. It will fail to run if the
user provides a string instead of a number.
}
