% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Construct File Path for a Specified Year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A four digit number representing a year in the format 'YYYY', numeric}
}
\value{
A character string representing a file path
}
\description{
`make_filename()` is a helper function used by `fars_read_years()` and `fars_map_state()`,
which constructs paths to annual FARS data files that corresponds to its argument and
follows a common naming pattern/convention.
}
\examples{
make_filename(2013)

yr <- 2013
file <- make_filename(yr)
}
