% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create filename according to Year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Indicating the year of interest. It should be entered as 4 digit
integer.}
}
\value{
a string specifying the filename
}
\description{
The function creates the file name according to the input argument. The
filename should correspond to the year of the FARS data and is in the format
of "accident_YYYY_.csv.bz2", with YYYY being the year of interest.
}
\examples{
\dontrun{make_filename(2013)}

}
